Working prototype of django react login
