+++ assets/js/auth.js
@@ -1,0 +1,41 @@
+module.exports = {
+    login: function(username, pass, cb) {
+        if (localStorage.token) {
+            if (cb) cb(true)
+            return
+        }
+        this.getToken(username, pass, (res) => {
+            if (res.authenticated) {
+                localStorage.token = res.token
+                if (cb) cb(true)
+            } else {
+                if (cb) cb(false)
+            }
+        })
+    },
+
+    logout: function() {
+        delete localStorage.token
+    },
+
+    loggedIn: function() {
+        return !!localStorage.token
+    },
+
+    getToken: function(username, pass, cb) {
+        $.ajax({
+            type: 'POST',
+            url: '/api-token-auth/',
+            data: {
+                username: username,
+                password: pass
+            },
+            success: function(res){
+                cb({
+                    authenticated: true,
+                    token: res.token
+                })
+            }
+        })
+    },
+}
