+++ assets/js/login.jsx
@@ -1,0 +1,29 @@
+var React = require('react')
+var auth = require('./auth')
+
+module.exports = React.createClass({
+    contextTypes: {
+        router: React.PropTypes.object.isRequired
+    },
+
+    handleSubmit: function(e) {
+        e.preventDefault()
+
+        var username = this.refs.username.value
+        var pass = this.refs.pass.value
+
+        auth.login(username, pass, (loggedIn) => {
+            this.context.router.replace('/react/')
+        })
+    },
+
+    render: function() {
+        return (
+            <form onSubmit={this.handleSubmit}>
+                <input type="text" placeholder="username" ref="username"/>
+                <input type="password" placeholder="password" ref="pass"/>
+                <input type="submit"/>
+            </form>
+        )
+    }
+})
