+++ betterself/templates/react.html
@@ -5,7 +5,7 @@
         <title>React Testing Title</title>
     </head>
     <body>
-        <div id="container"></div>
+        <div id="app"></div>
         {% render_bundle 'main' %}
         {# Just a simple react page to verify react works#}
     </body>
