+++ config/urls.py
@@ -19,6 +19,8 @@ urlpatterns = [
     url(r'^api/', include('apis.urls')),
     url(r'^analytics/', include('analytics.urls')),
     url(r'^react/$', TemplateView.as_view(template_name='react.html'), name='react'),
+    # curl -X POST -d "username=SOMETHING&password=SOMEPASSWORD" localhost:9000/api-token-auth/
+    url(r'^api-token-auth/', views.obtain_auth_token, name='api_token'),
 ]
 
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
@@ -31,7 +33,6 @@ if settings.DEBUG:
         url(r'^403/$', default_views.permission_denied, kwargs={'exception': Exception('Permission Denied')}),
         url(r'^404/$', default_views.page_not_found, kwargs={'exception': Exception('Page not Found')}),
         url(r'^500/$', default_views.server_error),
-        # curl -X POST -d "username=SOMETHING&password=SOMEPASSWORD" localhost:9000/api-token-auth/
-        url(r'^api-token-auth/', views.obtain_auth_token, name='api_token'),
+
     ]
     urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
