+++ webpack-stats.json
@@ -1,1 +1,1 @@
-{"status":"done","chunks":{"main":[{"name":"main-3310e4d2cf199792eacd.js","path":"/sites/betterself/assets/bundles/main-3310e4d2cf199792eacd.js"}]}}
+{"status":"done","chunks":{"main":[{"name":"main-1adb6432ecc48cb88cbf.js","path":"/sites/betterself/assets/bundles/main-1adb6432ecc48cb88cbf.js"}]}}
